refactor: move DSP related classes out of 'core' module
