+++ dsp/dsp-api/src/main/resources/dsp-2025-1-odrl-profile.jsonld
