+++ dsp/dsp-contract-negotiation/src/main/java/org/eclipse/dataspacetck/dsp/verification/cn/AbstractContractNegotiationTest.java
@@ -14,10 +14,10 @@
 
 package org.eclipse.dataspacetck.dsp.verification.cn;
 
-import org.eclipse.dataspacetck.core.api.verification.AbstractVerificationTest;
+import org.eclipse.dataspacetck.dsp.system.api.verification.AbstractVerificationTest;
 import org.junit.jupiter.api.BeforeAll;
 
-import static org.eclipse.dataspacetck.core.api.message.MessageSerializer.registerValidator;
+import static org.eclipse.dataspacetck.dsp.system.api.message.MessageSerializer.registerValidator;
 
 public class AbstractContractNegotiationTest extends AbstractVerificationTest {
