+++ dsp/dsp-system/src/main/java/org/eclipse/dataspacetck/dsp/system/client/cn/http/HttpConsumerNegotiationClientImpl.java
@@ -20,13 +20,13 @@ import org.eclipse.dataspacetck.dsp.system.client.cn.ConsumerNegotiationClient;
 import java.util.Map;
 
 import static java.lang.String.format;
-import static org.eclipse.dataspacetck.core.api.message.MessageSerializer.processJsonLd;
 import static org.eclipse.dataspacetck.dsp.system.api.http.HttpFunctions.getJson;
 import static org.eclipse.dataspacetck.dsp.system.api.http.HttpFunctions.postJson;
 import static org.eclipse.dataspacetck.dsp.system.api.message.DspConstants.DSPACE_PROPERTY_PROVIDER_PID_EXPANDED;
 import static org.eclipse.dataspacetck.dsp.system.api.message.DspConstants.DSPACE_PROPERTY_STATE_EXPANDED;
 import static org.eclipse.dataspacetck.dsp.system.api.message.DspConstants.TCK_PARTICIPANT_ID;
 import static org.eclipse.dataspacetck.dsp.system.api.message.JsonLdFunctions.stringIdProperty;
+import static org.eclipse.dataspacetck.dsp.system.api.message.MessageSerializer.processJsonLd;
 
 /**
  * Implementation of a {@link ConsumerNegotiationClient} that supports dispatch to a remote connector system via HTTP.
