+++ dsp/dsp-system/src/main/java/org/eclipse/dataspacetck/dsp/system/client/tp/http/HttpProviderTransferProcessClient.java
@@ -19,8 +19,8 @@ import org.eclipse.dataspacetck.dsp.system.client.tp.ProviderTransferProcessClie
 
 import java.util.Map;
 
-import static org.eclipse.dataspacetck.core.api.message.MessageSerializer.processJsonLd;
 import static org.eclipse.dataspacetck.dsp.system.api.http.HttpFunctions.postJson;
+import static org.eclipse.dataspacetck.dsp.system.api.message.MessageSerializer.processJsonLd;
 
 /**
  * Implementation of {@link ProviderTransferProcessClient} when running with remote connector
