+++ dsp/dsp-system/src/main/java/org/eclipse/dataspacetck/dsp/system/pipeline/ConsumerNegotiationPipelineImpl.java
@@ -30,12 +30,12 @@ import java.util.function.BiFunction;
 import java.util.function.Consumer;
 
 import static java.util.UUID.randomUUID;
-import static org.eclipse.dataspacetck.core.api.message.MessageSerializer.processJsonLd;
-import static org.eclipse.dataspacetck.core.api.message.MessageSerializer.serialize;
 import static org.eclipse.dataspacetck.dsp.system.api.message.DspConstants.DSPACE_NAMESPACE;
 import static org.eclipse.dataspacetck.dsp.system.api.message.DspConstants.DSPACE_PROPERTY_STATE_EXPANDED;
 import static org.eclipse.dataspacetck.dsp.system.api.message.DspConstants.TCK_PARTICIPANT_ID;
 import static org.eclipse.dataspacetck.dsp.system.api.message.JsonLdFunctions.stringIdProperty;
+import static org.eclipse.dataspacetck.dsp.system.api.message.MessageSerializer.processJsonLd;
+import static org.eclipse.dataspacetck.dsp.system.api.message.MessageSerializer.serialize;
 import static org.eclipse.dataspacetck.dsp.system.api.message.NegotiationFunctions.createAgreement;
 import static org.eclipse.dataspacetck.dsp.system.api.message.NegotiationFunctions.createFinalizedEvent;
 import static org.eclipse.dataspacetck.dsp.system.api.message.NegotiationFunctions.createOffer;
